+++ benchmark/benchmark.py
@@ -24,8 +24,8 @@ from dotenv import load_dotenv
 from plots import plot_refactoring
 from rich.console import Console
 
-from aider import models
-from aider.coders import Coder
+from aider import models, sendchat
+from aider.coders import Coder, base_coder
 from aider.dump import dump  # noqa: F401
 from aider.io import InputOutput
 
@@ -158,6 +158,9 @@ def main(
     dirnames: Optional[List[str]] = typer.Argument(None, help="Directory names"),
     graphs: bool = typer.Option(False, "--graphs", help="Generate graphs"),
     model: str = typer.Option("gpt-3.5-turbo", "--model", "-m", help="Model name"),
+    sleep: float = typer.Option(
+        0, "--sleep", help="Sleep seconds between tests when single threaded"
+    ),
     edit_format: str = typer.Option(None, "--edit-format", "-e", help="Edit format"),
     editor_model: str = typer.Option(None, "--editor-model", help="Editor model name"),
     editor_edit_format: str = typer.Option(None, "--editor-edit-format", help="Editor edit format"),
@@ -267,6 +270,11 @@ def main(
     if num_tests > 0:
         test_dnames = test_dnames[:num_tests]
 
+    # Don't give up when benchmarking
+    LONG_TIMEOUT = 24 * 60 * 60
+    sendchat.RETRY_TIMEOUT = LONG_TIMEOUT
+    base_coder.RETRY_TIMEOUT = LONG_TIMEOUT
+
     if threads == 1:
         all_results = []
         for testname in test_dnames:
@@ -285,10 +293,13 @@ def main(
                 editor_model,
                 editor_edit_format,
                 num_ctx,
+                sleep,
             )
 
             all_results.append(results)
             summarize_results(dirname)
+            if sleep:
+                time.sleep(sleep)
     else:
         run_test_threaded = lox.thread(threads)(run_test)
         for testname in test_dnames:
@@ -565,6 +576,7 @@ def run_test_real(
     editor_model,
     editor_edit_format,
     num_ctx=None,
+    sleep=0,
 ):
     if not os.path.isdir(testdir):
         print("Not a dir:", testdir)