+++ scripts/pip-compile.sh
@@ -5,6 +5,7 @@ set -e
 
 # First compile the base requirements
 pip-compile \
+    --allow-unsafe \
     requirements/requirements.in \
     --output-file=requirements.txt \
     $1
@@ -15,6 +16,7 @@ CONSTRAINTS="--constraint=requirements.txt"
 
 for SUFFIX in "${SUFFIXES[@]}"; do
     pip-compile \
+        --allow-unsafe \
         requirements/requirements-${SUFFIX}.in \
         --output-file=requirements/requirements-${SUFFIX}.txt \
         ${CONSTRAINTS} \